Have maxPasswordLength in XJKeyChain be a UInt32. Rename make_event_ical to show it's applescript nature.

+++ src/Carbon Wrappers/XJKeyChain.h
@@ -4,7 +4,7 @@
 
 @interface XJKeyChain : NSObject
 
-@property (nonatomic) NSUInteger maxPasswordLength;
+@property (nonatomic) UInt32 maxPasswordLength;
 
 + (instancetype)defaultKeyChain;
